+++ packages/cryptoplease/lib/bl/outgoing_transfers/create_outgoing_transfer_bloc/nft/event.dart
@@ -1,0 +1,24 @@
+part of 'bloc.dart';
+
+@freezed
+class NftCreateOutgoingTransferEvent with _$NftCreateOutgoingTransferEvent {
+  const factory NftCreateOutgoingTransferEvent.typeUpdated(
+    OutgoingTransferType transferType,
+  ) = TypeUpdated;
+
+  const factory NftCreateOutgoingTransferEvent.nftTransferCreated(
+      NonFungibleToken token) = NftTransferCreated;
+
+  const factory NftCreateOutgoingTransferEvent.recipientUpdated(
+      String address) = RecipientUpdated;
+
+  const factory NftCreateOutgoingTransferEvent.referenceUpdated(
+      String reference) = ReferenceUpdated;
+
+  const factory NftCreateOutgoingTransferEvent.memoUpdated(String memo) =
+      MemoUpdated;
+
+  const factory NftCreateOutgoingTransferEvent.cleared() = Cleared;
+
+  const factory NftCreateOutgoingTransferEvent.submitted() = Submitted;
+}
