+++ packages/cryptoplease/lib/bl/outgoing_transfers/create_outgoing_transfer_bloc/nft/state.dart
@@ -1,0 +1,24 @@
+part of 'bloc.dart';
+
+@freezed
+class NftCreateOutgoingTransferState with _$NftCreateOutgoingTransferState {
+  const factory NftCreateOutgoingTransferState({
+    required OutgoingTransferType transferType,
+    NonFungibleToken? nft,
+    Amount? maxFee,
+    String? recipientAddress,
+    String? memo,
+    String? reference,
+    @Default(FlowInitial<Exception, OutgoingTransferId>())
+        Flow<Exception, OutgoingTransferId> flow,
+  }) = _NftCreateOutgoingTransferState;
+
+  const NftCreateOutgoingTransferState._();
+
+  Amount get fee => calculateFee(
+        transferType,
+        nft!.address,
+      );
+
+  Token get token => nft!;
+}
