+++ packages/cryptoplease/lib/bl/tokens/token.dart
@@ -98,34 +98,20 @@ class SplToken extends Token {
           tags: tags,
           extensions: extensions,
         );
-
-  factory SplToken.nft({
-    required String address,
-    required String name,
-    required String logoURI,
-  }) =>
-      SplToken(
-        chainId: 0,
-        address: address,
-        symbol: '',
-        name: name,
-        decimals: 1,
-        logoURI: logoURI,
-        tags: const <String>[],
-        extensions: null,
-      );
 }
 
 class NonFungibleToken extends SplToken {
   const NonFungibleToken({
     required String address,
     required this.metadata,
+    String? logo,
   }) : super(
           chainId: currentChainId,
           address: address,
           symbol: '',
           name: '',
           decimals: 0,
+          logoURI: logo,
           tags: const <String>[],
         );
