+++ packages/cryptoplease/lib/presentation/screens/authenticated/send_flow/common/qr_scanner_screen/qr_scanner_screen.dart
