+++ packages/cryptoplease/lib/presentation/screens/authenticated/send_flow/fungible_token/confirm_screen/components/list_item.dart
