+++ packages/cryptoplease/lib/presentation/screens/authenticated/send_flow/fungible_token/enter_amount/component/token_list_dialog/search_box.dart
