+++ packages/cryptoplease/lib/presentation/screens/authenticated/send_flow/routes.dart
@@ -1,12 +1,15 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/bl/qr_scanner/qr_scanner_request.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/confirm_screen/confirm_screen.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/enter_address/enter_address_screen.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/enter_amount/enter_amount_screen.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/pick_recipient_type/pick_recipient_type_screen.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/qr_scanner_screen/qr_scanner_screen.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/send_nft_flow.dart';
-import 'package:cryptoplease/presentation/screens/authenticated/send_flow/send_token_flow.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/common/pick_recipient_type/pick_recipient_type_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/common/qr_scanner_screen/qr_scanner_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/fungible_token/confirm_screen/confirm_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/fungible_token/enter_address/enter_address_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/fungible_token/enter_amount/enter_amount_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/fungible_token/send_token_flow.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/non_fungible_token/confirm_screen/confirm_screen.dart';
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/non_fungible_token/enter_address/enter_address_screen.dart';
+
+import 'package:cryptoplease/presentation/screens/authenticated/send_flow/non_fungible_token/send_nft_flow.dart';
 
 const createPaymentFlowRoutes = [
   AutoRoute<void>(
@@ -14,9 +17,9 @@ const createPaymentFlowRoutes = [
     children: [
       AutoRoute<void>(path: '', page: PickRecipientTypeScreen),
       AutoRoute<QrScannerRequest>(page: QrScannerScreen),
-      AutoRoute<void>(page: EnterAddressScreen),
+      AutoRoute<void>(page: EnterFungibleTokenAddressScreen),
       AutoRoute<void>(page: EnterAmountScreen),
-      AutoRoute<void>(page: ConfirmScreen),
+      AutoRoute<void>(page: ConfirmFungibleTokenScreen),
     ],
   ),
   AutoRoute<void>(
@@ -24,8 +27,8 @@ const createPaymentFlowRoutes = [
     children: [
       AutoRoute<void>(path: '', page: PickRecipientTypeScreen),
       AutoRoute<QrScannerRequest>(page: QrScannerScreen),
-      AutoRoute<void>(page: EnterAddressScreen),
-      AutoRoute<void>(page: ConfirmScreen),
+      AutoRoute<void>(page: EnterNonFungibleTokenAddressScreen),
+      AutoRoute<void>(page: ConfirmNonFungibleTokenScreen),
     ],
   ),
 ];
